+++ src/Swap/test/Token.sol
@@ -2,7 +2,6 @@
 pragma solidity 0.8.11;
 
 contract Token {
-
     string public symbol;
     string public name;
     uint256 public decimals;
@@ -13,12 +12,17 @@ contract Token {
 
     event Transfer(address from, address to, uint256 value);
     event Approval(address owner, address spender, uint256 value);
-    event LogChangeVault(address indexed oldVault, address indexed newVault, uint indexed effectiveTime);
+    event LogChangeVault(
+        address indexed oldVault,
+        address indexed newVault,
+        uint256 indexed effectiveTime
+    );
 
     bytes32 public DOMAIN_SEPARATOR;
     // keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
-    bytes32 public constant PERMIT_TYPEHASH = 0x6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c9;
-    mapping(address => uint) public nonces;
+    bytes32 public constant PERMIT_TYPEHASH =
+        0x6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c9;
+    mapping(address => uint256) public nonces;
 
     address public anyswapRouter;
     address public pendingAnyswapRouter;
@@ -34,13 +38,15 @@ contract Token {
         name = _name;
         symbol = _symbol;
         decimals = _decimals;
-        uint chainId;
+        uint256 chainId;
         assembly {
             chainId := chainid()
         }
         DOMAIN_SEPARATOR = keccak256(
             abi.encode(
-                keccak256('EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)'),
+                keccak256(
+                    'EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)'
+                ),
                 keccak256(bytes(name)),
                 keccak256(bytes('1')),
                 chainId,
@@ -50,23 +56,41 @@ contract Token {
         _mint(msg.sender, 0);
     }
 
-    function approve(address _spender, uint256 _value) public returns (bool) {
+    function approve(address _spender, uint256 _value)
+        public
+        returns (bool)
+    {
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
         return true;
     }
 
-    function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external {
+    function permit(
+        address owner,
+        address spender,
+        uint256 value,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    )
+        external
+    {
         require(deadline >= block.timestamp, 'StableV1: EXPIRED');
         bytes32 digest = keccak256(
             abi.encodePacked(
                 '\x19\x01',
                 DOMAIN_SEPARATOR,
-                keccak256(abi.encode(PERMIT_TYPEHASH, owner, spender, value, nonces[owner]++, deadline))
+                keccak256(
+                    abi.encode(PERMIT_TYPEHASH, owner, spender, value, nonces[owner]++, deadline)
+                )
             )
         );
         address recoveredAddress = ecrecover(digest, v, r, s);
-        require(recoveredAddress != address(0) && recoveredAddress == owner, 'StableV1: INVALID_SIGNATURE');
+        require(
+            recoveredAddress != address(0) && recoveredAddress == owner,
+            'StableV1: INVALID_SIGNATURE'
+        );
         allowance[owner][spender] = value;
 
         emit Approval(owner, spender, value);
@@ -76,26 +100,25 @@ contract Token {
         return address(this);
     }
 
-    function balance(address account) external view returns (uint) {
+    function balance(address account) external view returns (uint256) {
         return balanceOf[account];
     }
 
-    function claimFees() external returns (uint, uint) {
-        return (0,0);
+    function claimFees() external returns (uint256, uint256) {
+        return (0, 0);
     }
 
-    function _mint(address _to, uint _amount) internal returns (bool) {
+    function _mint(address _to, uint256 _amount) internal returns (bool) {
         balanceOf[_to] += _amount;
         totalSupply += _amount;
         emit Transfer(address(0x0), _to, _amount);
         return true;
     }
 
-    function _transfer(
-        address _from,
-        address _to,
-        uint256 _value
-    ) internal returns (bool) {
+    function _transfer(address _from, address _to, uint256 _value)
+        internal
+        returns (bool)
+    {
         balanceOf[_from] -= _value;
         balanceOf[_to] += _value;
         emit Transfer(_from, _to, _value);
@@ -106,13 +129,12 @@ contract Token {
         return _transfer(msg.sender, _to, _value);
     }
 
-    function transferFrom(
-        address _from,
-        address _to,
-        uint256 _value
-    ) public returns (bool) {
+    function transferFrom(address _from, address _to, uint256 _value)
+        public
+        returns (bool)
+    {
         uint256 allowed_from = allowance[_from][msg.sender];
-        if (allowed_from != type(uint).max) {
+        if (allowed_from != type(uint256).max) {
             allowance[_from][msg.sender] -= _value;
         }
         return _transfer(_from, _to, _value);
